+++ NEWS.md
@@ -1,3 +1,10 @@
+# procs 1.0.5
+
+* Fixed bug on `proc_means()` when ordered factor used as class.
+* Updates to documentation.
+* Fixed bug on `proc_transpose()` when factor used on by parameter.
+* Added "dist" option to `proc_ttest()`.
+
 # procs 1.0.4
 
 * Added "as.character" parameter to `proc_sort()` to cast any factors in the